% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{EnsemblHg19GenePos}
\alias{EnsemblHg19GenePos}
\title{Ensembl hg19 coordinates of gene start and end base pair positions}
\format{
Each row in this data set is a unique pair of HGNC gene symbols and Ensembl IDs. Their chromosomes, start, end, and base pair midpoint positions are provided
}
\source{
Ensembl hg19 coordinates downloaded from the Biomart web tool.
}
\usage{
EnsemblHg19GenePos
}
\description{
Each row in this data set is a unique pair of HGNC gene symbols and Ensembl IDs. Their chromosomes, start, end, and base pair midpoint positions are provided.
}
\examples{
data(EnsemblHg19GenePos)
str(EnsemblHg19GenePos)
}
\keyword{datasets}
