% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{EnsemblHg38GenePos}
\alias{EnsemblHg38GenePos}
\title{Ensembl hg38 coordinates of gene start and end base pair positions}
\format{
Each row in this data set is a pair of HGNC gene symbols and Ensembl IDs. Their chromosomes, start, end, and base pair midpoint positions are provided
}
\source{
Ensembl hg38 coordinates downloaded from the Biomart web tool.
}
\usage{
EnsemblHg19GenePos
}
\description{
Each row in this data set is a pair of HGNC gene symbols and Ensembl IDs. Their chromosomes, start, end, and base pair midpoint positions are provided.
}
\examples{
data(EnsemblHg38GenePos)
str(EnsemblHg38GenePos)
}
\keyword{datasets}
