% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{gent}
\alias{gent}
\title{Gene-based association test (GenT)}
\usage{
gent(zs = NULL, LD, A = NULL, chisquares = NULL)
}
\arguments{
\item{zs}{vector of SNP-specific Z-statistics from GWAS whose indices correspond to those of \code{LD}}

\item{LD}{LD matrix whose row/column indices correspond to the indices of \code{zs}}

\item{chisquares}{if you instead give chi-square statistics (squared Z-statistics), supply them here and leave \code{zs} empty}
}
\value{
\itemize{
\item \code{pval}: P-value for testing H0: gene is not associated with trait.
\item \code{shape}: shape parameter of null (Gamma) distribution.
\item \code{rate}: rate parameter of null (Gamma) distribution.
\item \code{mu_h0}: expectation of null (Gamma) distribution.
\item \code{sigma2_h0}: variance of null (Gamma) distribution.
\item \code{mu_h1}: observed statistic used to test H0.
\item \code{sigma2_h1}: variance of observed statistic used to test H0.
}
}
\description{
This function performs a single gene-based association test.
}
\examples{
# Example for 5 SNPs
LD=cov2cor(rWishart(1,100,diag(5))[,,1])
z=c(mvnfast::rmvn(1,rep(0,5),diag(5)))
gent(z,LD)
}
