% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{gent_manhattan}
\alias{gent_manhattan}
\title{Gene-based test Manhattan plot}
\usage{
gent_manhattan(
  gentres,
  chromosome = "chr",
  gene_position = "gene_start",
  gene_pvalue = "pval",
  significance_threshold = 0.05/12727,
  plot_threshold = 1e-100,
  label_genes = FALSE,
  gene_label = "gene",
  ld_population = "EUR"
)
}
\arguments{
\item{gentres}{Gene-based association test statistic results for multiple genes (e.g., output of <gent/mugent/xgent>_genomewide())}

\item{chromosome}{Column name of gene chromosome in \code{gentres} data}

\item{gene_position}{Column name of gene base pair position (e.g., TSS, start position) in \code{gentres} data}

\item{gene_pvalue}{Column name of gene-based test P-values in  \code{gentres} data}

\item{significance_threshold}{Genome-wide significance threshold. Used to draw a reference line and for annotating genes if you set \code{label_genes=TRUE}}

\item{plot_threshold}{Lower P-value threshold at which to truncate gene-based test P-values.}

\item{label_genes}{If TRUE, clumped genes (see \code{gene_clump()}) will be annotated at their position}

\item{gene_label}{Column name of gene identifier (e.g., gene symbol) in \code{gentres} data}

\item{ld_population}{LD population, one of 'EUR','AFR','EAS','SAS', or 'AMR'}
}
\value{
A \code{ggplot2} object which is a Manhattan plot of gene-based association test P-values
}
\description{
This function generates a Manhattan plot of gene-based test statistics
}
\examples{
gent_results=gent_genomewide(gwas_data,50,'EUR','ld_directory')
gent_manhattan(gent_results)
}
