% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{gent_qq}
\alias{gent_qq}
\title{Gene-based test QQ plot}
\usage{
gent_qq(gentres, gene_pvalue = "pval")
}
\arguments{
\item{gentres}{Gene-based association test statistic results for multiple genes (e.g., output of <gent/mugent/xgent>_genomewide())}

\item{gene_pvalue}{Column name of gene-based test P-values in  \code{gentres} data}
}
\value{
A \code{ggplot2} object which is a QQ plot of gene-based association test P-values
}
\description{
This function generates a quantile-quantile plot of gene-based test statistics
}
\examples{
gent_results=gent_genomewide(gwas_data,50,'EUR','ld_directory')
gent_qq(gent_results)
}
