% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{ld}
\alias{ld}
\title{Return LD matrix for gene}
\usage{
ld(
  rsids,
  effect_alleles,
  ld_ref_file = NULL,
  writeable_directory = getwd(),
  plink_exec = "plink",
  verbose = T
)
}
\arguments{
\item{rsids}{vector of rsIDs for the SNPs you want to estimate LD between.}

\item{effect_alleles}{effect alleles of the rsIDs from GWAS.}

\item{ld_ref_file}{filename (without extension) of LD reference panel in PLINK (.bim/.bed/.fam) format.}

\item{writeable_directory}{a directory in which you have read-write access. Files will be created in this directory then deleted from it.}

\item{plink_exec}{command to execute PLINK from the command line, typically just 'plink'.}

\item{verbose}{should a statement about your effect alleles and those in the LD reference be printed to the console?}
}
\value{
\itemize{
\item \code{ld}: LD matrix. Rows and columns are in the original order provided to this function, filtered to only those present in the LD reference.
\item \code{signmat}: This is a vector of +1s and -1s which indicate if your effect alleles did match (+1) or did not match (-1) the effect alleles in the LD reference panel. Effect alleles must match for later inference. If they don't simply multiply your Z-statistics, for SNPs present in the LD reference panel (see row/column names of \code{ld}), by \code{signment} element-wise.
}
}
\description{
This function can externally use PLINK to estimate the matrix of LD correlations between a set of SNPs.
}
\examples{
# Example for 10 SNPs
rs=paste0('rs',1:10)
ea=sample(c('A','C','T','G'),10,replace=T)
ld_ref='~/1kg.v3/EUR'
wd=getwd()
ld(rs,ea,ld_ref,wd,plink_exec='plink',verbose=TRUE)
}
