% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{mugent_genomewide}
\alias{mugent_genomewide}
\title{Genome-wide multi-trait/ancestry gene-based association test (MuGenT)}
\usage{
mugent_genomewide(
  gwas_list,
  ld_population_list,
  ld_directory = "ld_directory",
  KbWindow = 50,
  snp_list = lapply(1:length(gwas_list), "rsid"),
  chromosome_list = lapply(1:length(gwas_list), "chr"),
  position_list = lapply(1:length(gwas_list), "position"),
  effect_allele_list = lapply(1:length(gwas_list), "effect_allele"),
  z_statistic_list = lapply(1:length(gwas_list), "z"),
  mugentpleio_alpha = 0.05/12727,
  index = NULL,
  verbose = TRUE,
  return_snp_gene_pairs = FALSE
)
}
\arguments{
\item{ld_population_list}{List of LD populations (see Github Wiki)}

\item{ld_directory}{Relative or absolute filepath to LD directory (see Github Wiki)}

\item{KbWindow}{Kilobase window used to assign SNPs to genes}

\item{snp_list}{List of column names of SNP rsIDs in GWAS datasets}

\item{chromosome_list}{List of column name of SNP chromosomes in GWAS datasets}

\item{position_list}{List of column name of SNP hg19 base pair positions in GWAS datasetes}

\item{effect_allele_list}{List of column name of SNP effect alleles in GWAS datasets}

\item{z_statistic_list}{List of column name of SNP Z-statistics in GWAS datasets}

\item{mugentpleio_alpha}{MuGenT-Pleio nominal Type I error rate (uncorrected significance threshold)}

\item{index}{Gene index file. see \code{data(EnsemblHg19GenePos)} for the expected format.}

\item{verbose}{TRUE if progress should be printed to the console, FALSE otherwise}

\item{return_snp_gene_pairs}{If TRUE, will return a chromosome-specific list of tested gene-specific SNP sets annotated by gene}

\item{gwas}{List of population/trait-specific GWAS summary statistics}
}
\value{
A list of dataframes with these components:
\itemize{
\itemize{
\code{$MuGenT}
\item \code{test}: Type of test (MuGenT)
\item \code{gene}: Gene symbol
\item \code{chr}: Chromosome of gene
\item \code{pval}: P-value testing the gene-based null hypothesis
\item \code{shape}: Shape parameter of the Gamma null distribution
\item \code{rate}: Rate parameter of the Gamma null distribution
\item \code{mu_h0}: Expected value of the gene-based test statistic under the null hypothesis
\item \code{sigma2_h0}: Variance of the gene-based test statistic under the null hypothesis
}
\itemize{
\code{$MuGenT-PH}
\item \code{test}: Type of test (MuGenT-PH)
\item \code{gene}: Gene symbol
\item \code{chr}: Chromosome of gene
\item \code{pval}: P-value testing the gene-based null hypothesis
\item \code{shape}: Shape parameter of the Gamma null distribution
\item \code{rate}: Rate parameter of the Gamma null distribution
\item \code{mu_h0}: Expected value of the gene-based test statistic under the null hypothesis
\item \code{sigma2_h0}: Variance of the gene-based test statistic under the null hypothesis
}
\itemize{
\code{$MuGenT-Pleiotropy}
\item \code{test}: Type of test (MuGenT-PH)
\item \code{gene}: Gene symbol
\item \code{chr}: Chromosome of gene
\item \code{result}: Test result: 'pleitoropy' for evidence of pleiotropy, 'no pleiotropy' otherwise
\item \code{adjusted_significance_quantile}: Quantile of the null Gamma distribution corrected for multiple testing
}
}
}
\description{
This function performs multi-trait/ancestry gene-based association testing genome-wide.
}
\examples{
# Example for COVID-19 GWAS data in EUR and AFR populations
mugent_genomewide(
  gwas_list = list(EUR=eur_gwas, AFR=afr_gwas),
  ld_population_list = list(EUR='EUR', AFR='AFR'),
  ld_directory = 'ld_matrices',
  KbWindow = 50,
  snp_list = list(EUR='rsid', AFR='rsid'),
  chromosome_list = list(EUR='#CHR', AFR='#CHR'),
  position_list = list(EUR='POS', AFR='POS'),
  effect_allele_list = list(EUR='ALT', AFR='ALT'),
  z_statistic_list = list(EUR='z', AFR='z'),
  verbose = TRUE.
  return_snp_gene_pairs = FALSE)
}
