% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{mugent_pleio}
\alias{mugent_pleio}
\title{Multi-ancestry joint gene-based associated test (MuGenT-Pleio)}
\usage{
mugent_pleio(Z, ldlist, alpha = 0.05)
}
\arguments{
\item{Z}{matrix of Z-statistics from GWAS. Rows are SNPs and columns are populations.}

\item{ldlist}{list of population-specific LD matrices whose ordering corresponds to the column ordering of \code{Z}}

\item{alpha}{the nominal (uncorrected) significance threshold for testing a single gene. If testing all genes genome-wide, this should be less than the threshold for testing a single gene.}
}
\value{
\itemize {
\item \code{result}: An indication if H0 (the gene is not association in all populations/with all traits) can be rejected. 'Pleiotropy' if it is rejected and 'no pleiotropy' otherwise.
\item \code{adjusted_significance_quantile}: The chi-square quantile of the adjusted nominal significance threshold for inferring H1 given the number of SNPs used and populations/traits tested.
}
}
\description{
This function performs a test of H1 that the gene is associated with in all populations/for all phenotypes.
}
\examples{
# Example for 5 SNPs and 2 populations
ldlist=rWishart(2,100,diag(5))
ldlist=lapply(1:2,function(h) cov2cor(ldlist[,,h]))
Z=t(mvnfast::rmvn(2,rep(0,5),diag(5)))
mugent_pleio(Z,ldlist)
}
