% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{mugent_sel}
\alias{mugent_sel}
\title{Pairwise test of only one population with a gene association (MuGenT-Sel)}
\usage{
mugent_sel(Z, ldlist, verbose = T)
}
\arguments{
\item{Z}{matrix of Z-statistics from GWAS. Rows are SNPs and columns are populations.}

\item{ldlist}{list of population-specific LD matrices whose ordering corresponds to the column ordering of \code{Z}}
}
\value{
This function returns a matrix of pairwise comparisons between populations/traits. Non-missing values in this matrix represent posterior probabilities that the gene is only associated with one of the two traits/populations.
}
\description{
This function returns the posterior probability that a gene is only associated with one population/trait in a pair, pairwise for all available populations/traits.
}
\examples{
# Example for 5 SNPs and 2 populations
ldlist=rWishart(2,100,diag(5))
ldlist=lapply(1:2,function(h) cov2cor(ldlist[,,h]))
Z=t(mvnfast::rmvn(2,rep(0,5),diag(5)))
mugent_sel(Z,ldlist)
}
