% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{multipop_anova}
\alias{multipop_anova}
\title{SNP-based ANOVA test for population heterogeneity}
\usage{
multipop_anova(effect_size_matrix, standard_error_matrix)
}
\arguments{
\item{effect_size_matrix}{Matrix of GWAS SNP effect sizes. Rows are SNPs. Columns are populations/phenotypes.}

\item{standard_error_matrix}{Matrix of GWAS standard errors corresponding to \code{effect_size_matrix}.}
}
\value{
matrix whose columns respectively are the ANOVA P-value and test statistics
}
\description{
This function performs a single ANOVA test for SNP across populations/phenotypes.
}
\examples{
# Example for 5 SNPs and 2 populations
z=t(mvnfast::rmvn(2,rep(0,5),diag(5)))
b=z/sqrt(30000)
multipop_anova(b,b/z)
}
