% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{wgent_genomewide}
\alias{wgent_genomewide}
\title{Genome-wide inverse MAF-weighted gene-based association test}
\usage{
wgent_genomewide(
  gwas,
  snp_weights = "weights",
  KbWindow = 50,
  ld_population = "EUR",
  ld_directory = "ld_directory",
  snp = "rsid",
  chromosome = "chr",
  position = "position",
  effect_allele = "effect_allele",
  effect_size = "beta",
  standard_error = "se",
  index = NULL,
  verbose = TRUE,
  return_snp_gene_pairs = FALSE
)
}
\arguments{
\item{gwas}{GWAS summary statistics}

\item{snp_weights}{Column name in \code{gwas} indicating SNP weights (e.g., square root of inverse of MAF)}

\item{KbWindow}{Kilobase window used to assign SNPs to genes}

\item{ld_population}{LD population, one of 'EUR', 'AFR', 'EAS', 'SAS', or 'AMR'}

\item{ld_directory}{Relative or absolute filepath to LD directory (see Github Wiki)}

\item{snp}{Column name of SNP rsID in GWAS data}

\item{chromosome}{Column name of SNP chromosome in GWAS data}

\item{position}{Column name of SNP hg19 base pair position in GWAS data}

\item{effect_allele}{Column name of SNP effect allele in GWAS data}

\item{effect_size}{Column name of SNP estimated effect size in GWAS data}

\item{standard_error}{Column name of SNP standard error in GWAS data}

\item{index}{Gene index file. see \code{data(EnsemblHg19GenePos)} for the expected format.}

\item{verbose}{TRUE if progress should be printed to the console, FALSE otherwise}

\item{return_snp_gene_pairs}{If TRUE, will return a chromosome-specific list of tested gene-specific SNP sets annotated by gene}
}
\value{
A dataframe with these components:
\itemize{
\item \code{pval}: P-value testing the gene-based null hypothesis
\item \code{shape}: Shape parameter of the Gamma null distribution
\item \code{rate}: Rate parameter of the Gamma null distribution
\item \code{mu_h0}: Expected value of the gene-based test statistic under the null hypothesis
\item \code{sigma2_h0}: Variance of the gene-based test statistic under the null hypothesis
\item \code{gene}: Gene symbol
\item \code{m}: Number of SNPs tested in gene-specific set
\item \code{chr}: Chromosome of gene
\item \code{gene_start}: Start position of gene (Ensembl hg19)
\item \code{window_start}: Start position of window in which SNPs were captured for this gene (hg19)
\item \code{gene_end}: End position of gene (Ensembl hg19)
\item \code{window_end}: End position of window in which SNPs were captured for this gene (hg19)
}
}
\description{
This function performs gene-based association testing genome-wide using weights.
}
\examples{
# Example for Alzheimer's disease GWAS data
gent_genomewide(
  gwas=ad_gwas,
  snp_weights='sqrt_inverse_mafs',
  ld_population='EUR',
  ld_directory='ld_matrices',
  KbWindow=50,
  snp='MarkerName',
  chromosome='Chromosome',
  position='Position',
  effect_allele='Effect_allele',
  effect_size='beta',
  standard_error='se',
  verbose=TRUE,
  return_snp_gene_pairs=FALSE)
}
